Flesh out lightRow
God there's like 4 different things I should be doing right now, but this is the only one of them I understand how to do at the moment, so it's the one I'm doing first.

Anyway, I added arguments and description to lightRow. I THINK that it might actually be useful for this program, but I guess we'll see.