Create version that almost works
Turns out the reason the last commit was so bad, is because I forgot that 0s in test_serial are (LED_ON), not 1s....

So, this version almost works. EXCEPT the bottom two rows of my 16x16 test don't work (which is definitely a hardware issue, I  think), and all the pixels above an enabled pixel also turn on. I have literally no idea what's causing that second thing, but it definitely feels like the kind of thing that would be caused by software rather than hardware.