Identify vertical bleed
So, the problem I was having with bleeding upwards (IE: turning on a pixel in row 3, will also create a slight bleed into that same column in rows 1 and 2. It's caused by inability to fully disable the column pins while the row SR is getting into place. This is... not GREAT, but honestly there are bigger fish to fry.

This is mostly just a commit to get this down in writing; no major changes to the software. Sorry folks!